+++ src/main/java/com/babzip/backend/global/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -2,18 +2,13 @@ package com.babzip.backend.global.oauth.handler;
 
 import com.babzip.backend.global.jwt.JwtHandler;
 import com.babzip.backend.global.jwt.JwtUserClaim;
-import com.babzip.backend.global.jwt.TokenProvider;
 import com.babzip.backend.global.oauth.service.OAuth2UserPrincipal;
-import com.babzip.backend.global.oauth.user.OAuth2Provider;
-import com.babzip.backend.global.token.entity.Token;
-import com.babzip.backend.user.domain.User;
+import com.babzip.backend.token.entity.Token;
 import com.babzip.backend.user.domain.UserRole;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
